consomation d'api edit et update ..
