+++ frontendecommerce/src/pages/dashboard/layout.tsx
@@ -16,8 +16,14 @@ const Layout: React.FC<DashboardProps> = ({ children }) => {
       <div className="bg-gray-50">
         <Header />
         <Mobile />
+        
         <Sidebar />
+       
         <Main>{children}</Main>
+       
+        
+  
+      
       </div>
     </>
   );
